ui-redesign: Signup
